+++ httpx/concurrency.py
@@ -14,9 +14,9 @@ import ssl
 import typing
 from types import TracebackType
 
-from .config import PoolLimits, TimeoutConfig
-from .exceptions import ConnectTimeout, PoolTimeout, ReadTimeout, WriteTimeout
-from .interfaces import (
+from ..config import PoolLimits, TimeoutConfig
+from ..exceptions import ConnectTimeout, PoolTimeout, ReadTimeout, WriteTimeout
+from ..interfaces import (
     BaseBackgroundManager,
     BasePoolSemaphore,
     BaseReader,
@@ -24,6 +24,7 @@ from .interfaces import (
     ConcurrencyBackend,
     Protocol,
 )
+from .utils import TimeoutFlag
 
 SSL_MONKEY_PATCH_APPLIED = False
 
@@ -47,38 +48,6 @@ def ssl_monkey_patch() -> None:
             _write(self, data)
 
     MonkeyPatch.write = _fixed_write
-
-
-class TimeoutFlag:
-    """
-    A timeout flag holds a state of either read-timeout or write-timeout mode.
-
-    We use this so that we can attempt both reads and writes concurrently, while
-    only enforcing timeouts in one direction.
-
-    During a request/response cycle we start in write-timeout mode.
-
-    Once we've sent a request fully, or once we start seeing a response,
-    then we switch to read-timeout mode instead.
-    """
-
-    def __init__(self) -> None:
-        self.raise_on_read_timeout = False
-        self.raise_on_write_timeout = True
-
-    def set_read_timeouts(self) -> None:
-        """
-        Set the flag to read-timeout mode.
-        """
-        self.raise_on_read_timeout = True
-        self.raise_on_write_timeout = False
-
-    def set_write_timeouts(self) -> None:
-        """
-        Set the flag to write-timeout mode.
-        """
-        self.raise_on_read_timeout = False
-        self.raise_on_write_timeout = True
 
 
 class Reader(BaseReader):
