+++ httpx/concurrency/utils.py
@@ -1,0 +1,30 @@
+class TimeoutFlag:
+    """
+    A timeout flag holds a state of either read-timeout or write-timeout mode.
+
+    We use this so that we can attempt both reads and writes concurrently, while
+    only enforcing timeouts in one direction.
+
+    During a request/response cycle we start in write-timeout mode.
+
+    Once we've sent a request fully, or once we start seeing a response,
+    then we switch to read-timeout mode instead.
+    """
+
+    def __init__(self) -> None:
+        self.raise_on_read_timeout = False
+        self.raise_on_write_timeout = True
+
+    def set_read_timeouts(self) -> None:
+        """
+        Set the flag to read-timeout mode.
+        """
+        self.raise_on_read_timeout = True
+        self.raise_on_write_timeout = False
+
+    def set_write_timeouts(self) -> None:
+        """
+        Set the flag to write-timeout mode.
+        """
+        self.raise_on_read_timeout = False
+        self.raise_on_write_timeout = True
