+++ httpx/dispatch/http11.py
@@ -2,7 +2,7 @@ import typing
 
 import h11
 
-from ..concurrency import TimeoutFlag
+from ..concurrency.utils import TimeoutFlag
 from ..config import TimeoutConfig, TimeoutTypes
 from ..interfaces import BaseReader, BaseWriter, ConcurrencyBackend
 from ..models import AsyncRequest, AsyncResponse
